Introduced the PUSHNIL, PUSHINT, PUSHDOUBLE, and PUSHSTRING instructions, in order to replace the PUSHCELL instruction. The latter don't require existing cells to be in memory in order for the VM to run and are thus more appropriate for when one day we will store the compiled G-code in a bytecode file. It also means that when printing the compiled output for debugging purposes, the same output should be produced every time, by virtue of the fact that the instruction arguments are not pointers.
PUSHGLOBAL has also been changed to accept the number of the built-in function
or supercombinator which needs to be pushed onto the stack, instead of a cell
pointer which it previously took (like PUSHCELL). This was done for the same
reason the other instructions were introduced.

Added the -g option to nreduce to just print the compiled G-code and exit.
The compilation order has been changed so that supercombinators appear before
the G-code corresponding to the built-in functions, so that the latter can
be excluded from the output shown with -g.

The JIT compiler hasn't been updated to reflect these changes yet and as such
native execution mode doesn't work currently. The G-code compiler is currently
being transitioned to a different scheme and there may be other instructions
added as part of this; the machine code generator will be updated after this
is done.