Additional optimizations to the compiler:
Where a supercombinator is being applied to the correct number of arguments, compile the
strict ones with needseval = true, and then push a reference to the supercombinator which
will skip the EVAL instructions at the start.
  
Do some eager evaluation in the case of CONS - it's actually cheaper than doing two MKAPs.

Eagerly execute builtin functions directly when we know that all the arguments have been
evaluated, instead of creating the intermediate graph structure. This is safe to do
because all the (currently existing) builtin functions are cheap, and since the arguments
are evaluated there's no chance of us jumping into code that we shouldn't.

When recursively calling C to deal with application nodes, pass in n+1 as the value for n
so that MKAP(n) will be used at the end instead of a whole sequence of MKAP(1)
instructions.