Initial support for arrays.
Two new cell types are added: TYPE_ARRAY and TYPE_AREF. Cells of type TYPE_ARRAY have
their first field set to a carray*, which contains the size and pointers to each of
the cells in the array as well as the "tail" representing the rest of the list that
is not stored in the array. Cells of type TYPE_AREF have their first field set to the
array cell, and their second set to an index into that array.

This array mechanism is purely an optimisation strategy which is does not change the
semantics of the language. From the programmer's point of view, array references are
indistinguishable from cons cells. Calling head with an array reference will return the
item at that position in the list, while calling tail will return a new array reference
pointing to the subsequent position in the array. However, if tail is called with a
reference that points to the last item in the array, the carray structure's tail field
will be returned instead. In this way, an array can represent a portion of a list, with
the rest residing in linked cons structures. This is particularly useful for "caching"
portions of infinite lists, or representing partially evaluated lists in a manner that
can be efficiently indexed but not requiring the entire list to be evaluated.

samples/arraylife.l is currently the only program which uses arrays. The functions
item and len have been modified to treat the case of array references differently,
and operate in O(1) time in the case where their parameter is an array vs. O(n) for
a linked list. The first time either of these is called with a given list, it will
convert a portion of the list (in the case of item) or the entire list (in the case of
len) into an array, and the first cons cell in the list will be replaced with an array
reference with index 0. Thus, the first call will be O(n) and all subsequent calls
will be O(1). This brings a huge performance benefit to programs which perform lots of
array accesses (such as life) and does so without requiring any explicit direction
from the programmer.

It is intended that item and len will become functions provided by the engine rather
than something that user programs are intended to implement. This is partly because
they are common functions that should form part of the standard library, but also because
they use a number of internal builtin functions (such as convertarray and arrayext) which
have side effects and are intended only to be used in special circumstances and should
thus not be accessible to normal user code.