Switched to a simpler (but faster) strictness analyser. This one is based on fairly straightforward analysis of the supercombinator bodies, rather than abstract interpretation. The old one had execution time that was exponential in the number of arguments to a supercombinator, and since it was not uncommon to see many arguments in lambda-lifted supercombinators this resulted in rather poor startup performance.
I think the old version may also have had some obscure cases where it would incorrectly
(i.e. unsafely) mark some arguments as being strict when this was actually not the
case.

This new version misses some cases, most notably higher order functions, which are
now assumed to be strict in none of their arguments. This may have a small performance
penalty in some situations but I think it's a worthwhile tradeoff to increase the
performance of the compiler.