Changed all of the built-in functions to take an array of cells as their parameter, rather than a stack object. This permits us in the future to use something other than a stack object for storing the arguments to be passed into a function.
One implication of this is that the built-in function has no way of accessing (or
decrementing) the stack count. This is now the responsbility of the caller, who
can determine the number of elements to pop from the stack based on the number of
arguments the function is declared to take. This should be one less than the number
of arguments, because one item is left on the stack representing the return value
of the function.