Changed the ALLOC instruction so that it creates HOLE cells instead of NIL. The former has a different meaning - a cell which does not yet have a value and is waiting to be filled in (it's used when constructing a graph).
Also added the REPLACE instruction which is similar to UPDATE, but instead of modifying
the graph simply changes an item in the stack to point to the specified value. This is
used during the printing loop where UPDATE was previously overwriting CONS cells that
it probably shouldn't have been. Now, the only types of cells which should ever be
overwritten are APPLICATION, FUNCTION, and HOLE.