Changed the strictness algorithm to handle letrec expressions. Now, in addition to checking for strict usage of supercombinator arguments, it also checks for strict usage of letrec variables. This means that when constructing the letrec, the expression can be evaluated directly rather than being delayed. Note that this is only safe to do if we can guarantee that the expression doesn't reference any other (yet to be constructed) letrec value graphs.
This modification also obviates the need to convert each supercombinator body to
a graph and back during the latter stages of compilation. Strictness analysis just
leaves the body as is (adding annotations only), and the G-code compiler doesn't need
to call graph_to_letrec() since the body will already be a letrec if necessary.