Changed the printing mechanism to use a supercombinator written in lambda calculus, rather than hard-coded G-code instructions included at the start of the compiled program.
The reason for this is to ensure that *all* code has a maximum stack size, i.e. we can
guarantee for each instruction that it will never go beyond a certain stack position.
This means that in the future we can allocate fixed sized frames and avoid runtime checks
for stack position, since for each instruction we can statically check what it will do
with the stack and detect violations before running the code (similar to Java's bytecode
verifier).

This gets rid of the PRINT and ISTYPE opcodes and adds the built-in "print" function.
nreduce.c has been modified to parse a small piece of "internal" source code before
the main file containing the logic for the printing mechanism.