Re-wrote the lambda lifting algorithm (again!) so that it now operates on trees (containing letrecs) instead of graphs. This is a cleaner way of doing it and fixes some cases that the previous algorithm couldn't handle (e.g. liftchain.test).
The stricness analysis algorithm has also been modified to deal with letrecs, so there is
no longer any stage in the compilation process which requires the source tree to be
represented as a graph. The letrec expressions that the G-code compiler sees are thus
those that were specified in the source file, rather than the reconstructed version used
previously which turned a graph back into a letrec.

One of the main advantages of this is that letrec expressions are kept at the proper
level of nesting, and that a letrec expression that appeared in a strict context in a
source file remains in a strict context when its compiled. Previously, inner letrecs
were "floated out" to the top level of the supercombinator body, which meant that it was
not possible to optimise the code using strictness information about them.

A few other algorithms e.g. the supercombinator instantiation function in the reduction
engine have also been simplified as a result of being able to deal with trees instead
of graphs.

Also made a number of changes to the pretty printing code to make the source layout
nicer.