Reorder letrec bindings so that variables that are used earlier are bound earlier. This allows for more efficient compilation, because we can avoid allocating temprary HOLE cells in many cases. For example the following code:
(let ((c (+ b 4))
      (b (+ a 3))
      (a (+ 1 2)))
     (* c 5)))

Will be transformed to:

let ((a (+ 1 2))
     (b (+ a 3))
     (c (+ b 4)))
    (* c 5) 

In the former case, c is evaluated, followed by b, then a. These must be delayed
evaluations (requiring temporary allocations) because each depends on later values.
In the latter case, we can evaluate each of the expressions directly without having
to allocate any memory to hold the delayed evaluation.