Fixed the reduction engine which broke with commit 310. This was related to a very obscure problem. The old code said:
  s->data[s->count-1-i] = stack_pop(s);

Prior to commit 310, stack_pop() was defined to return a void*, but in 310 it became
cell*. As a result of this change gcc (3.3.6) decided to evaluate the address of the LHS
*before* making the function call, instead of after. Since stack_pop() modifies s->count,
this the result was being written to the wrong memory location. I guess perhaps the
order of evaluation isn't defined by the C standard for these cases.

The solution was to change the code to the following:

  cell *val = stack_pop(s);
  s->data[s->count-1-i] = val;

so that we can ensure the function call definitely happens first.

Also fixed a (harmless) case where stack_pop() was being used on a non-cell* stack.