Use an unboxed representation for numbers. This means that instead of having each number occupy a cell in memory (which must be allocated and subsequently garbage collected), store the value of the numbers directly in the pointer.
Previously, each pointer occupied 32 bits (or sizeof(void*), but I've only tested with
32 bit platforms so far). Because we support double-precision floating points however,
and we need to distinguish between numbers and pointers, we now use 64 bits. The way
that we store a pointer is by storing 1s in the first 12 bits (indicating an "invalid"
double value, i.e. nan) and using the last 32 bits to store the pointer. Thus it is
easy to check of a particular 64 bit value should be interpreted as a number or a
pointer.

An implication of this change is that it is now necessary to use a different
representation for cells in memory, since each field is now twice the size. The
engine now uses a different data structure (rtvalue) at runtime instead of cell,
which is identical expect for the field size. There are two reasons for this:

1. There is a large amount of code which deals with the existing cell objects at compile
time, which expects each field to be 32 bits. Changing this is a fair bit of work.

2. For a long time I've wanted to store extra information in the cell structure at
compile time, such as which line of source it corresponds to, for error reporting
and other purposes. However this would have meant larger stuctures at runtime, leading
to reduced cache utilization. Keeping them separate means that these fields can be
added in the future without affecting runtime performance.

Both the G-machine and reduction engine have been modified to use the new pointer/rtvalue
scheme at runtime, however the compilation code remains the same. A number of macros
have been added to deal with pointer construction/extraction which this code makes heavy
use of.

While there is a small price for the use of 64-bit pointers and the extra cell size, this
is greatly offset by the decrease in the number of cells the memory allocator has to
deal with at run time. For the compute-intensive code in the mandelbrot program
(the mandel supercombinator) this means that *no* memory allocation is done at all. Most
code that does a lot of number crunching will see significant benefits from this change.
Although the performance increase (79% for bigmandelbrot) wasn't quite as good as I
was expecting, it should lead to greater gains once the machine code generator is working
again.