Cleaned up the array handling a bit. Instead of converting linked lists of cons cells to arrays at the time that len/item is called, do it as part of the tail operation.
When tail is called with a cons cell, and it notices that the result is also a cons
cell, it merges the two into an array and replaces the original cons cells with
references into that array. When tail is called with the last reference in an array,
and it notices that the array's tail is also a cons cell, it appends that item to
the array.

The effect of this is that simply traversing a list will convert it into an array,
without any specific intervention from the programmer. Now the only optimizations
that are needed at "user level" are when accessing the array. See arraylife.l for
an example of how to do this (the code in arraylife will eventually be moved into the
global set of functions).