Properly report runtime errors while showing the correct line number in the source file which contains the code that caused the error. This requires us to keep information during the compilaton for every cell about which file and lineno it corresponds to. This information is also recorded in the generated G-code, so for any given instruction it is possible to tell the line of code that instruction was generated from.
Also changed struct cell to include type-specific fields rathern than a generic
void *field1 and void *field2. The former was necessary when cells were used at runtime
and we had to conserve memory, but it's much more intuitive to have fields like name
and body and bindings, with the appropriate types. Cells are only used during the
compilation process so the increased memory overhead does not have any impact on runtime
performance.