Big code re-organization to complete the transition from a single "cell" data structure to different versions of it which are used at compile and runtime.
The structure that prior to this commit was known as cell is now called snode ("source
node"). This is no longer garbage collected, and thus these objects must be freed when
appropriate during the compilation process (having two collectors didn't make sense and
it is not necessary at compile time).

The recently added structure that was known as rtvalue has now been renamed to cell,
and its members changed from cmp1 and cmp2 to field1 and field2. Thus, the runtime code
is basically back to normal, except now all the compilation stuff has been separated
out, and we have 64 bit pointers in cells.