More code reorginisation. Major cleanups to the way that bytecode compilation is handled - instead of exposing the gprogram structure this stuff is handled internally and it is simply necessary for outside code to call compile() to turn a source object into bytecode.
Renamed gcode.[ch] to bytecode.[ch], and gmachine.c to interpreter.c. This is to
reflect the fact that we've deviated somewhat from the original G-machine structure
as described in the literature, so while the code is still based heavily on these
ideas it's not considered an implementation of the G machine as such. Also, the
new names are more immediately understable to anyone looking at the code who is
not familiar with the G machine concept.

Added a couple of extensions to the syntax - { and } can now be used as synonyms for
( and ), and the infix : operator can be used as an alternative to cons. This latter
improvement makes expressing lists a lot nicer.