More optimisations:
Got rid of the "count" field in the frame structure. This is no longer needed, as we
can statically determine the size of the stack for each instruction. The "expcount"
field of the instruction structure is used instead, making the execution model of the
interpreter more like a 3-address machine. It's still not true 3-address code however,
because there are still opcodes which treat the temporary data area for a frame like
a stack, such as MKCAP, MKFRAME, and BIF.

Tried to cut down on the number of actual work that needs to be done for each instruction:
- Changed the arithmetic operators in builtins.c to avoid the cost of a switch and two
  function calls
- Modified the error handling logic so that we don't have to perform an explicit check
  after each builtin function, by having the error handler change the address of the
  current frame to an instruction which has the ERROR opcode.
- #ifdefed some additional code so that when compiled with NDEBUG set, even less checks
  get done

Added the ability to explicitly specify strictness information for supercombinator
arguments and letrec expressions. I was a bit reluctant to do this as it taints the
"purity" of the language, but I decided the performance gains are probably worth it.

Also got profiling working again