Use a saner mechanism for handling outgoing and incoming messages.
For incoming messages, instead of maintaining a separate list of incoming messages
from each worker, we maintain a list containing messages from all workers. When the
interpreter wants to receive the next message, we just have to grab the head of this
list, rather than searching through all of the lists.

For outgoing messages, maintain a buffer for each destination worker containing
all of the data that we are waiting to send to that worker. This is in contrast
to the previous scheme of having each outgoing message stored in a separate object
and stored in the "outgoing" list for that destination. Now, the send buffer may
contain multiple outgoing messages.