Instead of creating worker objects in anticipation of clients that will connect in the future, don't create these initially, and instead only create them when the clients actually connect. This means that each active worker object represents a connection.
Because of this change, we can no longer just set the worker id during the 
initialisation process. When a client connects, we have to find out what worker id it
represents. Now when a connection is established, but the connecting side and the
accepting side send their ids to each other.

The doio() routine is now used to handle new connections and completion events for
previously initiated asynchronous connect() calls, in addition to the normal data
transfer that it was previously used for. process_received() has also been modified
to treat the first four bytes received a socket as the id of the corresponding worker.