Removed the id field from each worker. Now when a message is sent, instead of trying to locate the worker with that particular id, the task's logical id -> physical id map is consulted to determine which hostname, port number, and local task id the message should be sent to. The list of connected workers is then searched for that hostname/port combination, and the message sent via that channel.
The workerinfo struct has now been renamed to connection, and all of the code in worker.c
operates using these. This is because not all connections now are necessarily other
workers - some may be clients connected to the debug console, and in the future others
may be sockets that are opened by running tasks for I/O purposes.

master.c has been modified to use a cut down version of the workerinfo structure. It also
no longer uses a nodeinfo object, instead storing the list of workers directly.