Ditched the idea of using asynchronous I/O in favour of having a separate thread responsible for handling all I/O operations.
My original plan was that I would implement a simple user-space threading library which
allowed each running task to have it's own stack and execution context, and a SIGALRM
handler would implement timeslicing (like in the userthreads program in the examples
directory). However there were various complications with implementing this so I
decided instead to use pthreads. Howver, pthreads and signals don't really work well
together.

So now there is a separate thread started at the beginning of the program which is
responsible for handling all I/O. It uses a select loop to handle incoming data or
waiting for connections to be ready to receive pending data, and transfers the messages
across the network as appropriate. There wasn't really much change needed for this, the
main thing being that instead of the interpreter thread calling doio() directly, it
just handles execution and blocks whenever necessary using pthread_cond_wait().

The current implementation is a mixture of the shared memory-based messaging system
that was in the code until recently, and the new socket communications architecture.
The way in which they've now been combined allows the best of both worlds - we can
utilise multiple CPUs by having several execution threads running at once, and we can
also pass messages between the machines. Thus in a cluster of SMP machines we should be
able to make better use of the processors than the previous idea of running a separate
copy of the VM for each processor and only communicating via sockets.