Changed the way that tasks are created so that all initialization is done via messages instead of each node relying on the assumption that all other nodes have initialised and started with the same task as it has.
A manager thread is now present which handles all requests for creation of new tasks.
This thread runs on each node. It creates a "dummy task" which is not actually used
to execute bytecode, but just as a hook into the communications mechanism. Another
similar dummy task is used as a "launcher" process, which handles all the necessary
communication with the manager threads on the different nodes.

When a task is started, the following occurs:

1. The launcher thread is started and executes the start_task() function
2. All hostnames specified in the host list are resolved to their ip address and
   an array of task ids is created to hold the endpoint addresses of the manager
   threads. These use a standard task id defined by MANAGER_ID
3. The launcher sends a NEWTASK message to each manager, containing the pid, groupsize,
   and bytecode for the new task
4. The managers each create a new task and reply with the NEWTASKRESP message, containing
   the localid of the newly created task. Since the launcher already knows the ip and
   port for each manager, it can construct the list of task ids (the idmap) after
   receiving these responses.
5. The launcher sends an INITTASK message to the managers, containing the idmap that
   it just produced from the collation of responses from step 4 and the node ips and
   ports from step 2
6. Each manager replies with INITTASKRESP to indicate that it has received the idmap
7. Now the launcher knows that all tasks can safely begin execution, so it sends a
   STARTTASK message to each manager
8. The managers reply with STARTTASKRESP to indicate that their respective tasks
   are now running

Once this process is complete the task execution proceeds as before. In the near future
this will be used to allow multiple processes (i.e. groups of tasks) to be started,
and for a client tool to be written which performs this process against a set of running
VMs.

Note: The concepts of tasks and communicator should probably be separated so that it's
possible for the manager and launcher to be implemented without having to create an
actual task object. This will require quite a few code changes but the overall operation
should remain very similar.