Moved to an almost completely decentralised model for launching workers.
Now all the master does is log in to each node and start up nreduce in worker mode.
Each worker listens on it's port for incoming connections but does not immediately
initiate any connections, since it doesn't know what other nodes are going to be
running.

A client must connect to the debug console and issue a series of "open" commands to
initiate connections to other workers. The setupconns shell script does this for the
case of the nodes green1..green5 on the test cluster. After running this script, all
the nodes will connect to each other.

Also, there is no program that is run at startup. A client connected to the debug
console of one of the workers must use a "run" command to start a program. That worker
will compile the specified source file, and distribute it do the other workers, and
then the program will begin execution.

All workers can now be started as normal processes on each of the nodes without the
master being involved. This makes the master mode of the nreduce program largely
obsolete; it will probably be replaced with a shell script in the near future.

The other change to startup is that each node is no longer given the hostname/ip it
should consider it's "public" IP. Each worker needs to know that IP it is referred to
by the other workers, as this is used when setting the ID map. But a given machine
could potentially have multiple public IPs (in the case of multiple network interfaces)
so there isn't necessary a "right" one to use. To get around this problem, whenever
two workers establish a connection to each other, they send their peer the IP address
that they know that peer as. When the worker receives this information it considers that
as it's public IP.