Separated the notion of task and communication endpoint. Up until now, entities in the system other than tasks (notably the manager thread and the process launcher) needed to create dummy "task" objects so that they could communicate with each other over the network. This was ugly because it meant that there was lots of unnecessary stuff initialised (like the heap, idmap, global address table etc), and really it's best to keep task objects for things that are actually executing tasks.
Now, an endpoint object can be used by anything which wants to communicate with other
parts of the system. Each task has an endpoint, and message exchange between tasks occurs
as it did before using the tasks' endpoints. But the manager thread and process launcher
also have their own endpoints which they use to communicate with each other, and there
is no longer a need to create pseudo-tasks for this purpose.