Added cons? function to check if a value is a list or not
Support the syntax '\"' and '\'' for declaring single and double quote characters;
emacs seems to get confused about bracket nesting depths if the \ is omitted

Added the appendn and listn functions, which allow expressions in the form of
(appendn 3 list1 list2 list3) and (listn item1 item2 item3), avoiding the need for
deeply nested calls to append or cons. Makes code look a bit nicer.