Implemented partial evaluation. This is a technique which essentially executes parts of the program at compile time, and produces a modified version of the syntax tree in which all constant expressions are replaced with their values. It is invoked within an additional compilation step which runs the partial evaluator on each supercombinator. During reduction, any expression that cannot be reduced because it contains variables to which the values are not known is left as-is. A number of transformation rules are used on the graph to replace certain expressions with their equivalents, which in many cases can permit further reduction to occur.
The motivation behind partial evaluation is because of the large number of (usually
unnecessary) tests that the code generated by the XPath/XSLT compiler will need to
perform. For example, the XPath addition operator has to first check if both its
arguments are sequences of exactly one item, and then if the items are both numbers,
before it can actually add them together. In many cases where we know the type of
the arguments, we can execute these tests at compile time, greatly increasing the
runtime performance of the program.

The full process is too complex to describe here; I have started documenting it in
notes/partial_evaluation.lyx. Numerous regression tests have been added which also
serve as examples of the code transformations that the partial evaluator can make.

A tracing option has also been added to the reduction engine which produces a series
of "dot" files showing the graph at each step in the process. These can be rendered
and collated into a single pdf file using the "tracetopdf" script. You can view an
animation of the reduction process by stepping through the pdf file page by page.

The tracing option also causes a text-based version of the graph to be printed out
to the terminal at each step. This makes use of the new functionality in tosyntax.c
which translates from the runtime representation of the program graph to a syntax
tree, which had to be implemented to support partial evaluation.

Another new optimisation is "letrec sinking". This moves letrecs as far into the
syntax tree as possible, which increases the chance that they could be evaluated
as strict. While this was mainly necessary to ensure the syntax trees reconstructed
from the graphs modified during partial evaluator retained this optimal letrec
placing, it can also help for user-written code where the letrecs have not been
placed carefully. bigmandelbrot.l is an example where the placing makes a big
difference - if the letrec within "mandel" is moved to the top level it runs several
times slower.

Some other small fixes and changes have been made:
- UPDATE instructions generated by the bytecode compiler now set the correct stack
  location to evaluated
- An application of a letrec expression to one or more arguments now works, due
  to a fix in the application lifting process
- The old "trace mode" is now renamed to "compile info", and invoked with the -c
  option
- The help message for the nreduce executable has been cleaned up a bit
- Variables within a source file that start with _v now work (they previously clashed
  with the auto-generated names)
- The pretty-printer has been fixed up a bit