Stopped using cell flags for graph traversal, and instead use a pntrset to record which nodes have been visited. This avoids the cost of doing an additional two traversals within clear_graph() that was previously needed in order to clear the flags.
Re-wrote graph_replace() so that it does not need to do a separate traversal to
detect which nodes contain the value to be replaced. Instead, everything is done
within the one traversal carried out by graph_replace_r(). Arrays are now also
handled better here and on graph_to_syntax().

Re-wrote the pntrmap class to be more like pntrset, i.e. it allocates all of the
entries in the hashtable as one array, rather than using a separate call to malloc()
for each. It is also now used by graph_to_syntax(), in place of p2smapping objects
which were stored in an array rather than a hashtable, and thus inefficient.

Modified the letrec sinking algorithm so that only four traversals of the syntax
tree are needed, instead of O(n), where n = no. of letrec bindings. This was
necessary due to the large number of letrec bindings that arose in some of the graphs
produced by the partial evaluator.

Added the teststring and testarray built-in functions, so that tests can now get hold
of something which is guaranteed to be an array (with elemsize = 1 or 8,
respectively). This was needed because it was hard to "manufacture" an array from
user code, since they are normally supposed to be indistinguishable from normal
lists. These functions can be used to make sure the various graph traversal
algorithms handle arrays properly.

In partial evaluation mode, went back to instantiating each supercombinator once,
instead of each supercombinator *reference* once. The change made recently broke
the gridxslt outputed code. I may go back to this if I can figure out how to get it
working reliably.

apply_rules() is no longer called within reduce(), because it caused O(n^2)
graph traversals. Instead, run_partial() calls apply_rules() after reduction has
finished, looping around if necessary.