Changed graph_replace() so that it does not create partial copies of a set of application nodes. This was previously causing a large amount of application lifting to occur, because shared partial applications were assigned a letrec variable, and then applied to the remaining arguments.
Only apply the if rule if the size of the graph being reduced by the partial
evaluator is within a factor of 10 of the original size of the supercombinator body.
This avoids situations where many applications of the if rule occurred, each making
two copies of the graph, which lead to exponential growth.

Changed run_partial() so that it actually instantiates the supercombinator *before*
calling reduce(). This is done so that the original size of the graph can be
determined.

Removed the increment of sc->used when a supercombinator is instantiated at runtime.
This was unnecessary, since we already increment the value when the scref is created
inside instantiate_scomb_r()