Fixed a regression caused by the previous commit, which came about as a result of the change to the interpreter which incremented the current instruction (address) *before* processing the opcode, instead of afterwards.
During EVAL when we block a frame we want to make sure that the EVAL instruction will be
executed again when the frame is unblocked, in case the thing we eval'd returned a frame
or remote reference. Because the current instruction has now been incremented at this
point, we have to decrement it so that the instruction will be re-executed when control
returns to the current frame.