Automatically spark arguments of strict application nodes. This enables parallelism to be obtained without the programmer explicitly having to spark expressions.
Changed the fish delay to 250ms, and modified the delay mechanism to record the
*next* fish time, rather than the last. This allows us to add a random extra bit of
time to the delay, to try to avoid all processors sending fishes at the same time/rate.
Also, the fish is now sent to a random machine, rather than just the next one in the
task list.