Changed array_to_string() to properly handle cons cells, pointer arrays, and chains of arrays/conses. This should now cover all case where a valid string (list of numbers) is passed to array_to_string(), assuming that all elements in the list have been evaluated.
Also modified check_array_convert() to allow arrays to be converted into strings even
if their tail is non-nil, as long as it's not a frame.

Callers of array_to_string() must now check to the result to see if an invalid
(non-character) element was found in the list, and report an error of this occurs.