Changed the networking code so it's more generic. socketcomm is now called node, and all of the generic "node"-related stuff from network.c, worker.c, and task.c has been moved into the new node.c. The node.h header file also contains the relevant data structures. This code could potentially be used for other purposes than nreduce; the main reason for making it more generic at this point is to pave way for supporting generic socket connections from within ELC programs.
Endpoint localids are now assigned within node_add_endpoint(). This is so that the
appropriate locking that is necessary can be centralised, and it is easier to hand
out ids here than having several places in the code that do it. task_new() has been
modified so that it no longer takes a localid parameter; instead, it takes a reference
to the node, on which it calls node_add_endpoint().

task_kill() now works, at least during shutdown. It sets tsk->done to 1 and sends an
interrupt to the task. The task notices the done flag set and exits. Killing of tasks
at times other than shutdown is still not tested.

The client side of the distributed process creation logic has now been moved to
client.c.