Added a callback mechanism so interested parties can be notified of events that occur within the node. These include:
- Connection established
- Connection failed
- Connection accepted
- Connection closed
- I/O error when reading/writing connection
- System shutdown

Also added support for multiple listening sockets. Each of these is represented by
a listener object, which contains the port no., socket, and a callback function and
user pointer. This is for other things like listening sockets that are opened from ELC
programs (though not supported yet), and possibly a web server for providing a
monitoring page. Listeners can be registered on the node in a similar manner to
event callback functions.