Properly kill the launcher thread if it is running when the node wants to shut down
Handle the case where a task is killed before execute() starts, i.e. when it does not
have a thread associated with it. This can occur if node shutdown happens while
distributed process creation is in progress.

Added a logfile field to the node, and functions to add messages to the log. A the
moment it just prints the messages to stdout.