When accepting a connection from another node, use getsockname() to deterine the IP address on which the other node connected to us, rather than having the node send it as part of the data stream (which was an ugly solution). Now the only information apart from the welcome message that needs to be exchanged during the handshaking process is the other node's listen port.
Modified start_listening() so that if 0 is passed in as the port no. then the port
assigned by the OS will be returned in the additional parameter, outport. This is
useful for clients which need to act as a "node", but don't care what port they listen
on. In this case the port will be dynamically assigned.

When connecting to another node, assign the hostname field in the connection object
based on doing a reverse lookup on the ip, instead of using the string passed into
node_connect(). This gives uniformity to the names assigned by the client and server
end. If you pass in an IP address string to node_connect(), the connection will thus
have the hostname (if one exists for that IP).