When a task completes and execute() returns, the task is now removed from the list of running tasks on the node. This required changes to the way that tasks are killed as well, since there were some nasty issues related to thread synchronisation (e.g. if a task exits half way through being killed).
task_kill() has been renamed to task_kill_locked(), to signify that it must be called
with the node lock held. Rather than calling pthread_join() on the task's thread, it now
waits for the endpoint associated with the task to be closed (which happens right at the
end of the task destruction process). The task thread is now created within task_new(),
and immediately detached, so there is no need to store a reference to it or join/detach
it later. Because a task cannot task executing immediately however, it blocks on a
semaphore until its idmap is initialised.

Changed the default operation of nreduce to start up a node, run a single task on that
node, and then when the task has completed, shutd down the node and exit. This means
that whenever a task is being executed using the bytecode interpreter, there is *always*
a node available. This is a necessary prerequisite for the support of socket
connections within ELC programs.